Server-side stream interceptor
The server receives the metadata from client, telling which client ID
and caller arguments are wanted.
It will assign a new request ID for each request and wire up connection
streaming to the client.
Some options like which local logger to use and if we should caller are
passed as parameters when initializing the interceptor.