Implement client streaming interceptor side
The client initializes and pass to the clientID (from pid) and its
options like sending the callstack to the server.
The client will then examine each messages, try to see if this is a Log
format one and if so, print it with given logger.
Otherwise, it will forward it to the client directly.