Context idler for client stream
The context idler allow to intercept any calls resetting an idler
timeout.
If the client doesn’t receive any trace of server activity in term of
response or logs, the client will timeout and exits in error.

Even logs that are not displayed (debug logs from server, where the
client only print INFO or +) are taken into account resetting the timer.