Log each request and prints calls generically.
Use our connectionnotify interceptor in the daemon for this, and use the
context to know where to log it.
Consequently, change the order of the interceptors so that the logstream
initialization is always the first, and context is ready.
Adapt the tests to the new API and protect against nil pointers or
interfaces passed.