Decorellate showing caller from formatting
Separate it in an independant field so that the cli itself decide to
show or not the caller and format it.
We still register and send caller if any of local, remote or
forwarder asked for callstack information.
In that case, this is forwarded to all clients and only those who
requested it will decide to print them.
The code is cleaner under that logic.