Special case for dconf db files
dconf db files are binary files that are different on every mahcine and
the tests will always fail when executed on different machines.
dcfon db files are excluded from golden directories and from
comparison and we now test only that they exist after the execution of
dconf update through ApplyPolicy.

Co-authored-by: Didier Roche <didrocks@ubuntu.com>