Protect stdforwarder against file descriptor stealing
As we were using the destination for file descriptor to close our pipe,
that was all good as long as we know what the others packages are doing.
For instance, in tests, we capture stdout and stderr after, and so add a
pipe on top of the pipe.
We were thus closing the wrong pipe and can’t rely on dest, which was
creating a deadlock in tests. Capture what we captured over to ensure
we are closing the correct file.