Reorganize and fix test issue when running all tests with ./...
We got some errors like the following, when running all tests with
./...:
permission denied: call to polkit failed: The name org.freedesktop.PolicyKit1 was not provided by any .service files"
The theory is that the docker containers with polkitd hadn’t the time to
properly initalize, this was only when we were running all tests because
go test ./... will compile future tests binaries while running current
ones, and so, would slow down the computer triggering this issue.
Reformating TestMain to be the first thing we saw in the main package,
followed by helpers and relax the time for the container to start.