Condition machine initial refresh on sssd.conf
Only start machine GPO download on boot (blocking) if we have AD configured.
Instead of having a failure, we thus see directly that the unit failed
to start due to missing file.

We let the other units failing for logging in or refresh so that the
admin sees that there is something wrong in the configuration.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>