Depends on sssd offline status to try fetching the GPOs
- request to sssd in what status we are in
- if we are online, do the samba/ldap requests as we used to do, but
  with a contextual timeout.
Do not use gpolist status returned code for determining if we are
offline or not. Only base our state on sssd as this is what logged it in
or not.
We are doing the request everytime we want to download the GPO list, as
sssd will flip status based on network connectivity change.

Note that sssd doesn’t always flip the state when going back offline
having no network connected. In that case. we let gpo list script
failing (either exit 2 or timing out). This will be the case only for
refreshes, and the machine is already offline, but wrongly reported
online. Keep thus last known good state and report the failure in logs.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>