Don’t fail service startup if ad_server is empty
We will let the service to start, but will only fail the Updates() GRPC
calls. All the other calls are safe without any Active Directory Server.
The Server will be empty if we are offline and no server is set manually
in sssd.conf or in adsys options.

Factor out ad server discovery method We will reuse it when updating.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>