Use full length account name by default
Account name was truncated to 15 characters to be compatible with
Netbios limitations. Although this must not always be the case because
on non-Netbios AD the name is not truncated and thus the account is not
found.
We now try to full length account name, then use the truncated one if
there is no match.

Co-authored-by: Didier Roche <didrocks@ubuntu.com>