Handle CI/local dev linters
We need to user another profile for the dev environment and can’t reuse
the same list with --fast.
Also, we can’t provide a local workspace directory in vscode to point to
the correct file, hence this order -> default easy checker, CI -> point
the our checker list.
Unfortunately, that means a local run will need to point to the CI file
if we want to run all of them.