Change strategy for policykit priviledge escalation
Our previous approach was flawed: it was preventing any actions to
anyone not being an administrator.
Also, it was giving only auth_admin* permissions to admins.

We need to change the list of policykit administrators instead of the
action rules themselves.
For this, we need to keep an empty file if nothing happens, and only
consider the fact of disable local system administrators + set admins
from AD interaction.

We can easily disallow policykit administrator privilege by overriding
the AdminIdentities section, missing the previous system administrators
names.
Consequently, if we set additional adminitrators, we need to readd them
to the same list.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>