Support for appending strategy through different GPOs
Keys can be overridable, non overridable.
The append strategy means from a GPO standpoint that the further GPO value is
listed before closest GPO (and then, enforced GPO in reverse order).
This can be extended to support preprnd but it is implemented yet as there is
no real world cases.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>