Add different metaEnabled and metaDisabled keys.
We will need those keys for strategy enablement. It means that disabled
keys will have no empty values, but potentially a meta and strategy that
we will use later on.
Consequently, we now keep a disabled value in addition to the enabled
one for the container / basic keys.
We have 3 levels of overrides:
Meta key in expanded policy is the basic value.
It can be overriden by MetaEnabled or MetaDisabled to be more specific.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>