Allow scripts order file to be missing
We might not have shutdown but only startup scripts, or no logon but
logoff.
This would fail the whole system unit. We don’t want though that a
manual run fail it.
Consequently, if the script directory is ready, a flag allows to control
this behaviour.
Adapt tests to it.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>