Require a GPT.INI at the root of assets directory
We now always download the assets directory, to avoid the following
scenario when it was on computers refresh only:
1. One computer is booted (refresh machine policy and assets/scripts
   download is done)
2. The admin adds a new policy rules for user script
3. An user tries to log in before next computer refresh period
   (30 minutes) -> failure as the script doesn’t exist yet in local
   assets cache.
Many variations of this scenarios exists with updated assets/scripts.

Thus, we always try to download assets for both users and computers,
alongside gpos.

Support wrong assets directory and empty assets.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>