Add other user running scripts in integration tests
This helps ensuring that each user has its set of scripts and order file
in its own directory.
Add a new scripts section to an existing GPO. This impacts the default
user as well, but it means that we are now testing multiple repetitions
of the same scripts in the GPO.
We have to mock the chown calls to have the tests passing.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>