Basic type with values needs a marker to be disabled
Setting a DisabledValue on a key will never set it to **del, but will
file with the value.
If there is a value, still consider it as not being disabled. Then, we
add a specific marker if we want the value to be ignored and the key
marked as DISABLED.
Adapts and add tests for this.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>