Add the ability to merge multiple coverages
For adwatchd we will have a number of packages where tests must be run
with sudo on Linux, as indicated by the `ADSYS_SKIP_SUDO_TESTS`
environment variable.

As we want coverage for both the sudo and non-sudo runs, update the
coverage generation to concatenate the output of both runs.

We now write the "mode: set" line manually since it's only needed once
in the final coverage file.