Minor watcher improvements when updating dirs
- return an error if we are requested to watch an empty slice of
  directories
- log a message instead of returning an error if the `stopCmd` call
  fails: right now we only error if the watcher is already stopped - this
  shouldn't prevent us from starting the watcher just below

Both of these issues were observed in the autopkgtests. The first one is
due to filesystem shenanigans as logs show multiple events when the
config file is updated, so it could happen that we attempt to unmarshal
an empty file before the contents are updated.

The second one I could not fully pinpoint but I think the change makes
sense from an implementation perspective.