Fix loading policy content from uppercase folders
On some default GPOs created by Microsoft such as 'Default Domain
Policy', the object class directory names can be uppercase (e.g. MACHINE
or USER instead of Machine or User).

Currently we only check the capitalized version of the directory (e.g.
Machine), silently failing to parse any rules if we can't find a
directory that matches this name exactly.

To account for this case, check for both capitalized and uppercase
directories when parsing GPOs -- first the capitalized variant which is
most common, then the uppercase one. If none of the variants are found,
continue with the previous behavior of printing a debug message.

Fixes #346 / DEENG-303