Allow parsing policy entries with empty values
Even though we do not support this in our Ubuntu policies, it appears
that having policy entries with empty values is possible, as there are
Microsoft keys that present this behavior.

Currently, if we encounter such a policy entry we fail hard and avoid
parsing the rest of the policy, which could contain valid Ubuntu keys.

To be more lenient, consider this error to be a soft error. Store it on
the entry objects, and check it before applying the policies. Policy
entries that do not match our key prefix are already excluded, so we
just need to ensure our own policy entries don't have errors.

I've updated the tests, and added a filtered key with an empty value to
the `filtered` policy which fails with the previous code but passes now.

Closes #386 / DEENG-335