Lowercase target name when normalizing
AD accounts can login with any combination of upper/lowercase spelling
of their usernames, which is then used verbatim by adsys when applying
policies.

This means that logging in with `administrator@example.com` and
`Administrator@EXAMPLE.COM` will cache 2 sets of the same policies,
create 2 dconf profiles, and create 2 symlinks in `/run/adsys/krb5cc` to
the same (thankfully) Kerberos ticket. This is just a list of behaviors
that were easily observed, but there might be others as well.

After login, the username reported by `whoami`, and the username used by
adsys when running `adsysctl update` is lowercase, so it's safe to
assume that using a lowercase name will always work.

Update the target normalization function to also lowercase the target
name.

Closes #378 / DEENG-314