Update go generate to shell out to mkdir
In our long quest of debugging the "directory is a file" FTBFS issue on
adsys, we tried the following:

- adding `syscall.Sync` calls after we perform IO operations
- adding `sync` calls in the `rules` file after code generation
- adding sleeps after code generation (even large amounts like 30s)
- pre-creating all directories in the rules file using `mkdir` calls (we
  then started seeing files showing up as directories...)
- disabling parallelism in debuild (DEB_BUILD_OPTIONS=parallel=1)

Knowing that go can bypass glibc and use syscalls directly, this can
confuse fakeroot's filesystem hijacking, so opt for a "dumb" way of
solving things and just shell out to `mkdir -p` instead of calling
`os.MkdirAll` in our generators.

This appears to work, as I've kicked 10 successive builds on Launchpad
which all passed for every architecture. Compared with the current state
of things where we have at least a failing arch per build, this is way
better.