Refactoring and improving mount.go
Now the ApplyPolicy function in only supposed to distribute the entries
into their respective handler (ignoring keys that are not recognized)
and control the mutexes.

Split the writeMountsFile function into two functions, one to handle the
parsing of the entry value and one to safely write the file (.new first
and then replace the previous one) and fix the ownership.