Exporting ApplyUserPolicy for tests
applyUserPolicy is a private function. However, it makes more sense for
it to be listed in the mount_test.go file, as it tests a core part of
the mount package.

To do that, the function is now exported as "public" in the export_test
file, in order for it to be used in the package tests.