Fixing compare of golden files in TestPolicyUpdate and updating golden files
The comparison was being made between /run/HOSTNAME instead of
run/machine, which is the folder where we store the computer specific
scripts. This was making the golden files not properly reflect that we
should have.

The golden files have also been updated to properly reflect the run/
directory structure after the tests.