Implement partial removal and caching for apparmor
To handle the one profile -> many policies issue we make use of the
unprivileged `apparmor_parser -N` command which parses the input files
and returns a list of policy names which we can then pipe to
`apparmor_parser -R` for removal.

Profiles and policies under /etc/apparmor.d/adsys are cross-referenced
with the contents of /sys/kernel/security/apparmor/profiles to determine
which policies are actually loaded in the system. This is to ensure we
don't attempt to unload a policy under /etc/apparmor.d/adsys that
happens to not be loaded in the system for any reason.

Caching is built in into apparmor_parser, so we can leverage its
functionality. Adding -W when loading profiles will first attempt to
resolve profiles from cache if present, then write the updated cache to
/var/cache/adsys/apparmor.

DEENG-461 / DEENG-476