Update apparmor tests
* read-only root directory and no entries
This was failing before because we were removing the exec permissions on
the directory, so the os.Stat call in unloadAllRules would fail with a
"permission denied" error. The call succeeds now since the directory is
traversable.

* error on read-only (machine|.old|.new) directory
This actually fails in the same place, with the caveat that the nested
directories/files are able to be removed, whereas the top directory
cannot. Thus, only the golden files needed to be updated.