Add apparmor tests to policies manager
This involves a bit of argument restructuring because in order to have a
passing apparmor manager run we need control over both the
apparmor_parser command and the loaded policies file.

The current (and best) way to do it for now is to use functional
options. This will also aid us when updating the integration test to
take into account the apparmor manager.

I've repurposed the `loadedPoliciesFile` to be configurable via the
apparmorfs directory which is its parent directory and exposed the
configurable options to the policies manager.

For tests, the apparmor_parser command is stubbed to /bin/true for
convenience, and similar to the other managers we set the directory as
read-only in order to trigger a failure.