Rework skipping integration tests
With the addition of the Windows codebase we needed to skip certain
tests on Linux if we did not have root permissions. An
`ADSYS_SKIP_SUDO_TESTS` flag was added for this.

Additionally, some tests proved to be problematic when ran with `-race`
on Windows. Thus, we added `ADWATCHD_SKIP_INTEGRATION_TESTS` to account
for this.

Besides this, adsys already had `ADSYS_SKIP_INTEGRATION_TESTS` before
which was used to skip integration tests during Debian packaging and
autopkgtests.

This led to, among other things, tests failing if one ran go test ./...
without root privileges on Linux. To simplify the situation:
- automatically skip tests requiring root and log an appropriate message
- drop ADSYS_SKIP_SUDO_TESTS as it's accounted by the change above
- rename ADWATCHD_SKIP_INTEGRATION_TESTS to ADSYS_SKIP_INTEGRATION_TESTS
  in order to have a single environment variable controlling this

Closes #469 / DEENG-509