Implement apparmor user policy manager
The apparmor user policy manager works by laying down a mapping file
with the username at /etc/apparmor.d/adsys/users. It does not actually
run `apparmor_parser` on any of the files there, instead relying on the
machine policies to include the mappings at /etc/apparmor.d/adsys/users.

The admin can input a single filepath for the user entry, which we will
parse and wrap the contents in a hat/profile declaration block with the
current username, e.g.:

    ^username {
    <file content>
    }

After writing the file we will reload the machine policies for them to
take into account the user file changes.

If there are no changes in the file we will avoid needlessly calling
apparmor_parser.

When the admin disables or marks the user policy as not configured - we
will unload any policies pertaining to our specific user under
/etc/apparmor.d/adsys/machine, and remove the user mapping file at
/etc/apparmor.d/adsys/users.

DEENG-473 / DEENG-474