Refactor tracking and generating coverage files
Encapsulate the tracking methods fully in testutils. The API is
minimized and don’t rely on the consumer to pass the correct filename
anymore, as it’s internal.
Adapt python coverage to this new API.
Simplify the collect and make it (if needed) able to recollect after
first tracking.
Add many more comments on API documentation.