Update noParserOutput and destAlreadyExists behavior
For when noParserOutput is set to true, actually check if
apparmor_parser wrote something and fail if so. This surfaced some tests
where this occurred but we missed it.

For destAlreadyExists, rename this attribute to destsAlreadyExist and
handle both machine and user directories. This attribute will always
have 2 elements where the first refers to the existing machine
directory (if applicable), and the second to the existing users
directory (if applicable). If only one of them need to be set, the other
will have an empty string as a value.