Add package tests for apparmor user policy manager
These work by mirroring an actual binaries/roles/mappings setup, with
the pam_binaries and pam_roles in the machine subdirectory, and the user
mappings in the users subdirectory.

Having them properly linked up will allow us to run apparmor_parser -N
on the machine path in tests and get a list of valid user mappings.