Rework policiesFromFiles command execution
Previously we captured stdout and stderr together which could mess up
the output if apparmor_parser printed on both streams.

This can print harmless messages which don't cause the command to fail,
e.g. on my Arch machine:

    Cache read/write disabled: interface file missing. (Kernel needs
    AppArmor 2.4 compatibility patch.)

Capture stdout and stderr separately, and if we got something printed on
stderr log it to the user without failing the execution.