Properly clean up apparmor user policy file
When we have no user/machine entries configured, but there's a leftover
user profile in /etc/apparmor.d/users, we would return early from the
cleanup process because of the missing machine directory, leaving the
user profile intact.

Fix this by attempting to remove the user policy file either way, and
add a test to confirm the behavior.