Retry flaky Windows test
Wasn't able to pinpoint the exact reason why only this specific test
would fail with "context deadline exceeded" from time to time. I
increased the context timeout, added sleeps literally everywhere and it
still occurred.

Extract it from the table test (also gives us a chance to clean up most
of the variables there which were used for the sole purpose of this
specific test) and reimplment it in a separate test that retries the
`app.Run()` part a couple of times in case of errors.

To be even more on the safe side, if we get an unexpected error from
starting the service we will surface it and still fail the test.

Closes #470 / DEENG-510