Fixing issues with the mount unit generation.
The mount units where being created equally, despite the mount protocol
requested. This wasn't going to work, since different protocols require
different What values from the unit and the Type= was missing from the
unit template.

Since systemd.mount's basically become a call to the `mount` command,
the code had to be updated in order to account for the require syntax.