Get krb5cc using kinit for winbind backend
Unlike sssd, winbind doesn't store the Kerberos cached credentials for
the host anywhere, so we have to request this ourselves.

The simplest way to do this is to interact with the system keytab and
obtain the ticket for the principal using the `kinit` command.

To be able to implement this in the correct location, the backend
interface now also returns an error for the `HostKrb5CCNAME` function,
even though not all backends might make use of it.

DEENG-489