Add a bogus cargo-checksum.json file to please dh-cargo
For now, dh-cargo assumes that the sources it builds is a crate from
crates.io that has been processed by debcargo. Part of this assumption
is the presence of this file.

The good thing is, we'll be able to re-use it for something else later
on, as there are other checksum-related issues in our journey!