packaging: Build, test and install the Rust code
Sadly, we cannot fully rely on dh-cargo for everything as there are a
couple of issues that simply don't work well with it. Note that the
blame doesn't solely lie on dh-cargo or the custom wrapper, the upstream
cargo doesn't really help us either by refusing to install from
workspace Cargo.toml.

Because of that, we have to do quite a few things semi-manually.