Rust packaging (#549)
* Create a global Rust workspace at the top-level dir

This allows us to run cargo commands from the top-level, which will make
it much easier to use dh-cargo and friends.

* Add a bogus cargo-checksum.json file to please dh-cargo

For now, dh-cargo assumes that the sources it builds is a crate from
crates.io that has been processed by debcargo. Part of this assumption
is the presence of this file.

The good thing is, we'll be able to re-use it for something else later
on, as there are other checksum-related issues in our journey!

* During package build, skip Rust tests ran from Golang

We want to use the packaging Rust infrastructure to run them, it's much
easier than making the golang code aware of wrappers and such.

* Vendor in the Rust dependencies at clean time

This includes a prompt from dh-cargo to update (or rather create) the required
XS-Vendored-Sources-Rust field. That field is used to quickly find out
which package is built using crates with known vulnerabilities, and it's
checked at build-time as well.

I haven't included said field in the commit because:

1. It's going out of date pretty quickly, so versioning it without the
   vendored sources doesn't make much sense
2. It's a fairly ugly line in the d/control file:

> XS-Vendored-Sources-Rust: android_system_properties@0.1.5, anyhow@1.0.66, autocfg@1.1.0, bitflags@1.3.2, bumpalo@3.11.1, cc@1.0.77, cfg-expr@0.11.0, cfg-if@1.0.0, chrono@0.4.23, codespan-reporting@0.11.1, core-foundation-sys@0.8.3, cxx-build@1.0.82, cxx@1.0.82, cxxbridge-flags@1.0.82, cxxbridge-macro@1.0.82, futures-channel@0.3.25, futures-core@0.3.25, futures-executor@0.3.25, futures-io@0.3.25, futures-macro@0.3.25, futures-task@0.3.25, futures-util@0.3.25, gio-sys@0.16.3, gio@0.16.3, glib-macros@0.16.3, glib-sys@0.16.3, glib@0.16.4, gobject-sys@0.16.3, heck@0.4.0, iana-time-zone-haiku@0.1.1, iana-time-zone@0.1.53, itoa@1.0.4, js-sys@0.3.60, libc@0.2.137, link-cplusplus@1.0.7, log@0.4.17, num-integer@0.1.45, num-traits@0.2.15, once_cell@1.16.0, pin-project-lite@0.2.9, pin-utils@0.1.0, pkg-config@0.3.26, proc-macro-crate@1.2.1, proc-macro-error-attr@1.0.4, proc-macro-error@1.0.4, proc-macro2@1.0.47, quote@1.0.21, ryu@1.0.11, scratch@1.0.2, serde@1.0.148, serde_derive@1.0.148, serde_json@1.0.89, slab@0.4.7, smallvec@1.10.0, syn@1.0.104, system-deps@6.0.3, termcolor@1.1.3, thiserror-impl@1.0.37, thiserror@1.0.37, time@0.1.45, toml@0.5.9, unicode-ident@1.0.5, unicode-width@0.1.10, version-compare@0.1.1, version_check@0.9.4, wasi@0.10.0+wasi-snapshot-preview1, wasm-bindgen-backend@0.2.83, wasm-bindgen-macro-support@0.2.83, wasm-bindgen-macro@0.2.83, wasm-bindgen-shared@0.2.83, wasm-bindgen@0.2.83, winapi-i686-pc-windows-gnu@0.4.0, winapi-util@0.1.5, winapi-x86_64-pc-windows-gnu@0.4.0, winapi@0.3.9

V2: * Fix the bash syntax in the clean rule
    * Tweak the wording of the comment below the rule

* packaging: Build, test and install the Rust code

Sadly, we cannot fully rely on dh-cargo for everything as there are a
couple of issues that simply don't work well with it. Note that the
blame doesn't solely lie on dh-cargo or the custom wrapper, the upstream
cargo doesn't really help us either by refusing to install from
workspace Cargo.toml.

Because of that, we have to do quite a few things semi-manually.

* gitignore: ignore vendored Rust code

* Rust: Do the "configure" step manually to be compatible with Jammy