Add apparmor as a build dependency
For apparmor tests where we unload policies, we rely on the actual
output of apparmor_parser -N, as it doesn't require root and doesn't
affect the system in any way. This binary is provided by the apparmor
package which may not always be installed on the building machine.

Make sure it's always installed by adding it to the build dependencies.