Add winbind as an alternative dependency for adsys
In case the user already has sssd or winbind installed, do not require
them to install the other one as well. Prioritize sssd since that's the
main backend we expect users to use.

Do the same for krb5-user which provides the kinit command which is
currently only used when the backend is set to winbind.