Add proxy manager that sets environment proxy conf
Implement an API that will be used by a future proxy policy manager and
a CLI program to set the proxy configuration on the system.

Currently it only supports system-wide proxy configuration via an
environment file under `/etc/environment.d`, but in the future it will
also configure APT and GSettings proxy settings.

DEENG-553