Replacing os.WriteFile after init in the config tests
Viper relies on fsnotify to trigger a config update. However, the
fsnotify docs state that the Write event is also triggered by truncating
the file. This means that the update callback was being called too soon
and we were unmarshaling the file contents before they were properly
written.

To avoid this problem, the os.WriteFile call had to be replaced and now
we manually open the file, write to it and then close it to properly
flush the changes and then have the callback triggered.