Adding guideline docs to the policy managers
When creating the mount policy manager, we realized that we needed a
defined way of handling errors in the managers. Which errors should
prevent authentication? Which errors should only be warned to the user?

After some discussion, we defined an expected general behavior that
all managers should have and this must documented on the general
manager, defined in manager.go

Along with the general manager docs, this also adds guideline docs to
each specific policy manager.

Co-authored-by: Gabriel Nagy <GabrielNagy@users.noreply.github.com>