Update packaging with new checksum strategy
Our previous strategy of copying a fake checksum file to the problematic
crate was unsustainable, since dependencies can change (either directly
or indirectly) when updating our code (or our dependencies version) and
new crates with .a files could emerge.

To fix this, a script was created to parse each vendored dep changelog
to see if there's any .a file listed and, if so, replaces the dep
changelog with a more generic one, using only the general crate checksum
instead of the default per-file one.