Warn when Pro-only rules are configured
If the user has configured rules that are only eligible to Ubuntu Pro
subscribers, issue a warning in addition to filtering them out.

I've opted for this approach instead of listing the types in the
`applied` command output as the latter can change if the machine is
unenrolled from Pro after applying the Pro-only rules for example.

Additionally, as a user (and developer) I'd expect to be warned as early
as possible of any of my set rules not being applied, namely during the
execution of the `update` command. I've experienced this numerous times
when adding new managers and forgetting to remove the filtering for
them.

Fixes #422 / DEENG-481