Handle case mismatches in GPT.INI file name
Windows filesystems are case-insensitive, but things can get hairy when
we parse the assets in Linux land, which for the most part is
case-sensitive.

GPT files are always uppercase (both filename and extension) when
created by Windows, but we cannot fully rely on this as we instruct
users to create their own GPT.INI files for the assets directories. As
such, upper/lowercase combinations might crop up, such as `GPT.ini`,
`gpt.ini`, etc.

Up until now we've always considered `GPT.INI` to be the "canonical"
path, but we can be a bit more defensive by walking the local directory
and doing a case-insensitive comparison on each entry, stopping when we
find a matching GPT.INI file.

This is not such a visible bug because our default behavior is to
redownload the assets if we encounter any issues when parsing the local
GPT.INI file, so we have the side effect of always redownloading assets
even if there are no changes.

Fixes #667 / UDENG-227