Ensure empty state for dconf policy manager
To aid in the usage of the purge command, unify the empty state of the
dconf manager similar to our other managers.

For user policies, we can safely remove the DB/profile if we have no
entries to apply. For machine policies we cannot do this, as we do not
know if there are other user policies that might depend on it. Instead,
we handle purging the machine DBs/profiles in our prerm purge script.

This required a few changes in tests that expected profile data to still
exist when applying a GPO with 0 entries.

Fixes #678 / UDENG-221