Rework krb5cc handling logic
Following the Heimdal issues surfaced after the samba 4.16.0 release we
have to rethink the krb5cc handling logic with the following in mind:
- we can no longer use symlinks as Kerberos tickets
- the ticket has to be owned by the running user (in our case root)

This is not related to a change in samba itself, but to the Kerberos
implementation vendored in samba (Heimdal Kerberos). Between samba
4.15.0 and 4.16.0 there has been a massive update in the vendored code,
pulling in around 10 years worth of changes[1]. Among these changes we
were able to find information confirming the 2 issues above, and we
concluded they are not considered to be bugs in the upstream
implementation[2] (and related issue [3]).

Another important thing to note is that the issues above do not manifest
if samba is compiled and linked against the MIT Kerberos library.
However, MIT Kerberos appears to still be considered experimental (as of
today), and neither Ubuntu or other distros are compiling samba with MIT
Kerberos.

As such, we have to work around these "issues" within adsys. Instead of
tracking the ticket's lifetime using a symlink, we rely on the following
approach:

* still track original ticket with a symlink
* copy the ticket as a root-owned ticket
* track refresh with timestamps on the symlink itself.

Ticket copies will be stored under KRB5CACHEDIR, whereas symlinks will
reside under KRB5CACHEDIR/tracking.

Closes UDENG-289

[1] https://github.com/samba-team/samba/commit/40b65c840e03bd5eb7f3b02fe80144650c63c005
[2] https://github.com/heimdal/heimdal/commit/2a565482f491621fe8a906c990aa890be153a906
[3] https://github.com/heimdal/heimdal/issues/306