Fix policy update offline integration tests
I believe that with the previous commit we've uncovered a small bug in
the krb5cc symlink logic. More specifically, in these tests we were
creating symlinks to a nonexistent path (ccache_OFFLINE) due to the sssd
configuration.

Because the presence of the source file wasn't enforced in
`ensureKrb5CCName` (filepath.Abs doesn't ensure the path actually
exists), this went unnoticed until now. This was only a problem in the
offline tests which didn't go to our samdb mock, which would have caught
the dangling symlink issue.

Now that we do the additonal copy action the failure becomes visible.
The fix is simple, we just have to create the cached sssd ticket using
the appropriate domain name.