Update mock logic for gvfs_anonymous_error
In Rust, gio repeatedly called AskPassword when the first access was
denied, so we had to get creative in the code and in the mocks.

Using gio through CGo appears to not suffer from the same behavior, so
we need to update the way that the mode "gvfs_anonymous_error" causes a
failure in the mock.