Updating pam module to properly set DCONF_PROFILE
Our pam module wasn't parsing sssd.conf to check for the
default_domain_suffix configuration value when setting the DCONF_PROFILE
variable. This means that, for cases where the user authenticates
without an explicit domain, the DCONF_PROFILE would be set to the
username only, instead of user@domain.

By parsing the config value, now we are able to set the environment
properly regardless of the username configuration.